djangofsm/blogpost/api/v1/serializers.py
from rest_framework import serializers

from blogpost.models import BlogPost


class BlogPostSerializer(serializers.ModelSerializer):
    class Meta:
        model = BlogPost
        fields = ["id", "title", "state"]
