pyproject.toml
[tool.poetry]
name = "django-fsm-test"
version = "0.1.0"
description = ""
authors = ["Franz Andrei Cruspero <franz.cruspero@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
djangorestframework = "^3.13.1"
django-fsm = "^2.8.0"
django-fsm-log = "^2.0.1"
django-fsm-admin = "^1.2.4"
black = "^21.12b0"
drf-yasg = "^1.20.0"
graphviz = "^0.19.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
